displays options wrapper component, faster renders
