+++ electron/app/containers/SamplesContainer.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useRef } from "react";
-import { useRecoilState, useRecoilValue } from "recoil";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
 import styled from "styled-components";
 
 import { Grid, Sticky } from "semantic-ui-react";
@@ -27,7 +27,8 @@ const Root = styled.div`
   }
 `;
 
-const SamplesContainer = (props) => {
+const DisplayOptionsWrapper = (props) => {
+  const { containerRef, sidebarRef, stickyHeaderRef, displayProps } = props;
   const {
     activeTags,
     activeLabels,
@@ -35,40 +36,22 @@ const SamplesContainer = (props) => {
     setActiveTags,
     setActiveLabels,
     setActiveOther,
-    labelData,
-  } = props.displayProps;
-
-  const [showSidebar, setShowSidebar] = useRecoilState(atoms.sidebarVisible);
-  const [sidebarHeight, setSidebarHeight] = useState("unset");
-  const [stuck, setStuck] = useState(false);
-  const datasetName = useRecoilValue(selectors.datasetName);
-  const numSamples = useRecoilValue(selectors.numSamples);
+  } = displayProps;
+  const labelSampleCounts = useRecoilValue(selectors.labelSampleCounts);
+  const colorMapping = useRecoilValue(selectors.labelColorMapping);
   const tagNames = useRecoilValue(selectors.tagNames);
   const tagSampleCounts = useRecoilValue(selectors.tagSampleCounts);
+  const filters = useRecoilValue(selectors.labelFilters);
+  const setModalFilters = useSetRecoilState(selectors.modalLabelFilters);
+  const [sidebarHeight, setSidebarHeight] = useState("unset");
   const fieldSchema = useRecoilValue(selectors.fieldSchema);
   const labelNames = useRecoilValue(selectors.labelNames);
   const labelTypes = useRecoilValue(selectors.labelTypes);
-  const labelSampleCounts = useRecoilValue(selectors.labelSampleCounts);
-  const colorMapping = useRecoilValue(selectors.labelColorMapping);
 
-  const containerRef = useRef();
-  const stickyHeaderRef = useRef();
-  const sidebarRef = useRef();
-
-  const labelNameGroups = {
-    labels: [],
-    scalars: [],
-    unsupported: [],
-  };
-  for (const name of labelNames) {
-    if (VALID_LABEL_TYPES.includes(labelTypes[name])) {
-      labelNameGroups.labels.push({ name, type: labelTypes[name] });
-    } else if (VALID_SCALAR_TYPES.includes(fieldSchema[name])) {
-      labelNameGroups.scalars.push({ name });
-    } else {
-      labelNameGroups.unsupported.push({ name });
-    }
-  }
+  useEffect(() => {
+    setModalFilters(filters);
+  }, [filters]);
+  let headerHeight = 0;
 
   const getDisplayOptions = (values, counts, selected) => {
     return [...values].sort().map(({ name, type }) => ({
@@ -84,12 +67,9 @@ const SamplesContainer = (props) => {
       [entry.name]: entry.selected,
     }));
   };
-
-  let headerHeight = 0;
   if (stickyHeaderRef.current && stickyHeaderRef.current.stickyRect) {
     headerHeight = stickyHeaderRef.current.stickyRect.height;
   }
-
   const updateSidebarHeight = () => {
     if (sidebarRef.current) {
       setSidebarHeight(
@@ -100,16 +80,82 @@ const SamplesContainer = (props) => {
   useResizeHandler(updateSidebarHeight, [sidebarRef.current]);
   useScrollHandler(updateSidebarHeight, [sidebarRef.current]);
   useEffect(updateSidebarHeight, []);
+  const labelNameGroups = {
+    labels: [],
+    scalars: [],
+    unsupported: [],
+  };
+  for (const name of labelNames) {
+    if (VALID_LABEL_TYPES.includes(labelTypes[name])) {
+      labelNameGroups.labels.push({ name, type: labelTypes[name] });
+    } else if (VALID_SCALAR_TYPES.includes(fieldSchema[name])) {
+      labelNameGroups.scalars.push({ name });
+    } else {
+      labelNameGroups.unsupported.push({ name });
+    }
+  }
+
+  return (
+    <Grid.Column className="sidebar-column">
+      <Sticky
+        context={containerRef}
+        offset={headerHeight}
+        style={{ height: "100%" }}
+        styleElement={{ height: "100%" }}
+      >
+        <DisplayOptionsSidebar
+          colorMapping={colorMapping}
+          tags={getDisplayOptions(
+            tagNames.map((t) => ({ name: t })),
+            tagSampleCounts,
+            activeTags
+          )}
+          labels={getDisplayOptions(
+            labelNameGroups.labels,
+            labelSampleCounts,
+            activeLabels
+          )}
+          onSelectTag={handleSetDisplayOption(activeTags, setActiveTags)}
+          onSelectLabel={handleSetDisplayOption(activeLabels, setActiveLabels)}
+          scalars={getDisplayOptions(
+            labelNameGroups.scalars,
+            labelSampleCounts,
+            activeOther
+          )}
+          onSelectScalar={handleSetDisplayOption(activeOther, setActiveOther)}
+          unsupported={getDisplayOptions(
+            labelNameGroups.unsupported,
+            labelSampleCounts,
+            activeLabels
+          )}
+          style={{
+            maxHeight: sidebarHeight,
+            overflowY: "auto",
+            overflowX: "hidden",
+            paddingRight: 25,
+            marginRight: -25,
+            scrollbarWidth: "thin",
+          }}
+          ref={sidebarRef}
+        />
+      </Sticky>
+    </Grid.Column>
+  );
+};
+
+const SamplesContainer = (props) => {
+  const [showSidebar, setShowSidebar] = useRecoilState(atoms.sidebarVisible);
+  const datasetName = useRecoilValue(selectors.datasetName);
+  const numSamples = useRecoilValue(selectors.numSamples);
+
+  const containerRef = useRef();
+  const stickyHeaderRef = useRef();
+  const sidebarRef = useRef();
 
   return (
     <Root ref={containerRef} showSidebar={showSidebar}>
       <VerticalSpacer opaque height={5} />
-      <Sticky
-        ref={stickyHeaderRef}
-        context={containerRef}
-        onStick={() => setStuck(true)}
-        onUnstick={() => setStuck(false)}
-      >
+      <Sticky ref={stickyHeaderRef} context={containerRef}>
         <ViewBar />
         <VerticalSpacer opaque height={5} />
         <ImageContainerHeader
@@ -122,56 +168,12 @@ const SamplesContainer = (props) => {
       </Sticky>
       <Grid>
         {showSidebar ? (
-          <Grid.Column className="sidebar-column">
-            <Sticky
-              context={containerRef}
-              offset={headerHeight}
-              style={{ height: "100%" }}
-              styleElement={{ height: "100%" }}
-            >
-              <DisplayOptionsSidebar
-                colorMapping={colorMapping}
-                tags={getDisplayOptions(
-                  tagNames.map((t) => ({ name: t })),
-                  tagSampleCounts,
-                  activeTags
-                )}
-                labels={getDisplayOptions(
-                  labelNameGroups.labels,
-                  labelSampleCounts,
-                  activeLabels
-                )}
-                onSelectTag={handleSetDisplayOption(activeTags, setActiveTags)}
-                onSelectLabel={handleSetDisplayOption(
-                  activeLabels,
-                  setActiveLabels
-                )}
-                scalars={getDisplayOptions(
-                  labelNameGroups.scalars,
-                  labelSampleCounts,
-                  activeOther
-                )}
-                onSelectScalar={handleSetDisplayOption(
-                  activeOther,
-                  setActiveOther
-                )}
-                unsupported={getDisplayOptions(
-                  labelNameGroups.unsupported,
-                  labelSampleCounts,
-                  activeLabels
-                )}
-                style={{
-                  maxHeight: sidebarHeight,
-                  overflowY: "auto",
-                  overflowX: "hidden",
-                  paddingRight: 25,
-                  marginRight: -25,
-                  scrollbarWidth: "thin",
-                }}
-                ref={sidebarRef}
-              />
-            </Sticky>
-          </Grid.Column>
+          <DisplayOptionsWrapper
+            sidebarRef={sidebarRef}
+            stickyHeaderRef={stickyHeaderRef}
+            containerRef={containerRef}
+            {...props}
+          />
         ) : null}
         <Grid.Column className="content-column">
           <Samples {...props} />
